Kafka POC: containerize producers and provision Grafana
- Containerized Alerts and Telemetry demo producers and moved sample payloads under producer/app
- Added Grafana datasource/dashboard provisioning for zero-manual setup
- Replace hardcoded datasource UIDs with name-based references for portability
- Updated docker-compose stacks for readiness (healthchecks + depends_on: service_healthy) and resolved port/container-name collisions
- Updated Vector configs and dashboards (Telemetry dashboard cleanup; Alerts remap transform rename)
- Refreshed Kafka READMEs with Docker Compose instructions, endpoints, and guidance for disabling demo producer/local Kafka in real deployments
