--- Kafka/Alerts/grafana/alert-logs-dashboard-251024.json
@@ -1,159 +1,0 @@
-{
-  "annotations": {
-    "list": [
-      {
-        "builtIn": 1,
-        "datasource": {
-          "type": "grafana",
-          "uid": "-- Grafana --"
-        },
-        "enable": true,
-        "hide": true,
-        "iconColor": "rgba(0, 211, 255, 1)",
-        "name": "Annotations & Alerts",
-        "type": "dashboard"
-      }
-    ]
-  },
-  "editable": true,
-  "fiscalYearStartMonth": 0,
-  "graphTooltip": 0,
-  "id": 1,
-  "links": [],
-  "panels": [
-    {
-      "datasource": {
-        "type": "victoriametrics-logs-datasource",
-        "uid": "ff20etquxem0wf"
-      },
-      "fieldConfig": {
-        "defaults": {
-          "custom": {
-            "align": "auto",
-            "cellOptions": {
-              "type": "auto"
-            },
-            "inspect": false
-          },
-          "mappings": [],
-          "thresholds": {
-            "mode": "absolute",
-            "steps": [
-              {
-                "color": "green",
-                "value": 0
-              },
-              {
-                "color": "red",
-                "value": 80
-              }
-            ]
-          }
-        },
-        "overrides": []
-      },
-      "gridPos": {
-        "h": 14,
-        "w": 23,
-        "x": 0,
-        "y": 0
-      },
-      "id": 1,
-      "options": {
-        "cellHeight": "sm",
-        "footer": {
-          "countRows": false,
-          "fields": "",
-          "reducer": [
-            "sum"
-          ],
-          "show": false
-        },
-        "showHeader": true
-      },
-      "pluginVersion": "12.1.1",
-      "targets": [
-        {
-          "datasource": {
-            "type": "victoriametrics-logs-datasource",
-            "uid": "ff20etquxem0wf"
-          },
-          "editorMode": "code",
-          "expr": "system_identifier:~\"$system_id\"",
-          "queryType": "instant",
-          "refId": "A"
-        }
-      ],
-      "title": "Logs",
-      "transformations": [
-        {
-          "id": "extractFields",
-          "options": {
-            "delimiter": ",",
-            "keepTime": false,
-            "replace": false,
-            "source": "labels"
-          }
-        },
-        {
-          "id": "filterFieldsByName",
-          "options": {
-            "include": {
-              "names": [
-                "Time",
-                "Line",
-                "alert_id",
-                "eemi_message_id",
-                "is_acknowledged",
-                "severity",
-                "system_identifier",
-                "updated_timestamp"
-              ]
-            }
-          }
-        }
-      ],
-      "type": "table"
-    }
-  ],
-  "preload": false,
-  "schemaVersion": 41,
-  "tags": [],
-  "templating": {
-    "list": [
-      {
-        "current": {
-          "text": "6SXV903",
-          "value": "6SXV903"
-        },
-        "datasource": {
-          "type": "victoriametrics-logs-datasource",
-          "uid": "ff20etquxem0wf"
-        },
-        "definition": "",
-        "includeAll": true,
-        "name": "system_id",
-        "options": [],
-        "query": {
-          "field": "system_identifier",
-          "limit": 0,
-          "query": "",
-          "refId": "VictoriaLogsVariableQueryEditor-VariableQuery",
-          "type": "fieldValue"
-        },
-        "refresh": 1,
-        "regex": "",
-        "type": "query"
-      }
-    ]
-  },
-  "time": {
-    "from": "now-6h",
-    "to": "now"
-  },
-  "timepicker": {},
-  "timezone": "browser",
-  "title": "New dashboard",
-  "uid": "8921eb72-fa0a-487d-b38f-e1978dc42dcf",
-  "version": 3
-}
