+++ Kafka/Alerts/grafana/provisioning/datasources/victorialogs.yaml
@@ -1,0 +1,13 @@
+apiVersion: 1
+
+# Grafana datasource provisioning
+# This file auto-creates the VictoriaLogs datasource at Grafana startup.
+# It avoids manual datasource setup in the UI.
+datasources:
+  - name: VictoriaLogs
+    type: victoriametrics-logs-datasource
+    access: proxy
+    # Use the compose service name. From inside the Grafana container, localhost would be Grafana itself.
+    url: http://victorialogs:9428
+    isDefault: true
+    editable: true
