--- Kafka/Alerts/prometheus/prometheus.yml
@@ -1,7 +1,0 @@
-global:
-  scrape_interval: 10s
-
-scrape_configs:
-  - job_name: 'vector'
-    static_configs:
-      - targets: ['vector:8686']
