--- Kafka/Alerts/victoria-logs.yml
@@ -1,6 +1,0 @@
-version: "2"
-
-apiVersion: "v1"
-
-retention: "30d"
-logLevel: "info"
