--- Kafka/Metrics/docker-compose.yml
@@ -1,47 +1,0 @@
-version: "3.8"
-
-services:
-  kafka:
-    image: bitnami/kafka:latest
-    container_name: kafka
-    ports:
-      - 9092:9092     # Internal (for containers like Vector)
-      - 29092:29092   # External (for your host Python script)
-    environment:
-      - KAFKA_CFG_NODE_ID=1
-      - KAFKA_CFG_PROCESS_ROLES=controller,broker
-      - KAFKA_CFG_CONTROLLER_QUORUM_VOTERS=1@kafka:9093
-      - KAFKA_CFG_LISTENERS=PLAINTEXT_INTERNAL://:9092,PLAINTEXT_EXTERNAL://:29092,CONTROLLER://:9093
-      - KAFKA_CFG_ADVERTISED_LISTENERS=PLAINTEXT_INTERNAL://kafka:9092,PLAINTEXT_EXTERNAL://localhost:29092
-      - KAFKA_CFG_LISTENER_SECURITY_PROTOCOL_MAP=PLAINTEXT_INTERNAL:PLAINTEXT,PLAINTEXT_EXTERNAL:PLAINTEXT,CONTROLLER:PLAINTEXT
-      - KAFKA_CFG_CONTROLLER_LISTENER_NAMES=CONTROLLER
-      - KAFKA_CFG_INTER_BROKER_LISTENER_NAME=PLAINTEXT_INTERNAL
-      - ALLOW_PLAINTEXT_LISTENER=yes
-
-  vector:
-    image: timberio/vector:0.41.0-debian
-    container_name: vector
-    volumes:
-      - ./vector/vector.yaml:/etc/vector/vector.yaml
-    ports:
-      - 8686:8686
-    depends_on:
-      - kafka
-    command: ["--config", "/etc/vector/vector.yaml"]
-
-  victoriametrics:
-    image: victoriametrics/victoria-metrics
-    ports:
-      - 8428:8428
-
-  grafana:
-    image: grafana/grafana:latest
-    ports:
-      - 3000:3000
-    depends_on:
-      - victoriametrics
-    volumes:
-      - grafana-storage:/var/lib/grafana
-
-volumes:
-  grafana-storage:
