+++ Kafka/Telemetry/producer/requirements.txt
@@ -1,0 +1,2 @@
+kafka-python==2.0.2
+six==1.16.0
